Read extended description please.
This is the new file that is replacing writeFile.cpp. The purpose of this program is to translate ASCII maps into "array-compatible maps." Not all ASCII maps/sprites/characters, etc. follow the basic N x M shape that is needed for creating a simple array. Depending on the image size, this can be troublesome, as from my experience, the easiest way to print complex ASCII graphics is to use arrays, and run print statements through loops. Doing ASCII art within your .cpp program is tiresome, and very inconvenient. Every char you put towards your art needs 4x the work, as you are constantly surrounding it by "' char '," to make it array-compatible. This program fixes this. All you need is to do your ASCII art in a .txt file, which I call the raw file. After your art is finished, you pass the raw file through this program, create a refined file (which for user purposes isn't that useful, its just a buffer for the program to run its magic) and a final file, which is created on your computer, that is in perfect format for you to just copy paste it into your program. Now you should have an array compatible ASCII graphic.